move anghammarad function to digest file
